Fixed missing forward slash in routes, updated imports
